extract models from integration test
