+++ backend/api/tests/models/mod.rs
@@ -1,0 +1,2 @@
+mod project_details;
+pub use project_details::*;
